(tmp) Replace `context` handling DnD with Redux
