+++ frontend/src/components/boards/TaskBoard/TaskList.tsx
@@ -1,15 +1,16 @@
-import React, { useContext, useState } from 'react';
+import React, { useState } from 'react';
 
 import { useDrop } from 'react-dnd';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import { Card, CardActions, CardContent, Grid, Chip } from '@material-ui/core';
 
 import * as Model from 'models';
-import { useAppSelector } from 'utils/hooks';
+import { draggableItem, DragItem } from 'utils/dnd';
+import { useAppDispatch, useAppSelector } from 'utils/hooks';
+import { moveCard } from 'store/slices';
 import { LabeledSelect, ScrolledDiv } from 'templates';
-import { ListCardHeader, TaskCard } from '.';
 import { ButtonToAddTask } from '..';
-import { DragContext, DragItem, draggableItem } from '../DragContext';
+import { ListCardHeader, TaskCard } from '.';
 
 const borderWidth = '2px';
 const useStyles = makeStyles((theme: Theme) =>
@@ -56,9 +57,10 @@ const TaskList: React.FC<TaskListProps> = (props) => {
   const { list, listIndex } = props;
   const classes = useStyles();
   const selectedId = useAppSelector((state) => state.boards.infoBox.data?.id);
+  const dispatch = useAppDispatch();
   const [filterValue, setfilterValue] = useState<FilterName>(cardFilter.ALL);
-  const { dragDispatch } = useContext(DragContext);
 
+  /** リスト間のカードの移動を司る */
   const [, drop] = useDrop({
     accept: draggableItem.card,
     hover: (item: DragItem) => {
@@ -67,19 +69,19 @@ const TaskList: React.FC<TaskListProps> = (props) => {
       const dragIndex = item.index;
       const hoverIndex = 0;
 
+      // 位置不変の場合
       if (dragListIndex === hoverListIndex) return;
 
       const boardId = list.boardId;
-      dragDispatch({
-        type: 'MOVE_CARD',
-        payload: {
+      dispatch(
+        moveCard({
           dragListIndex,
           hoverListIndex,
           dragIndex,
           hoverIndex,
           boardId,
-        },
-      });
+        })
+      );
 
       item.index = hoverIndex;
       item.listIndex = hoverListIndex;
@@ -99,12 +101,8 @@ const TaskList: React.FC<TaskListProps> = (props) => {
     setfilterValue(event.target.value as FilterName); // unknown型から変換
   };
 
-  const handleDragEnd = () => {
-    // API request
-  };
-
   return (
-    <Card ref={drop} className={rootClass} onDragEnd={handleDragEnd}>
+    <Card ref={drop} className={rootClass}>
       <div className='listWrapper'>
         <ListCardHeader list={list} />
