+++ frontend/src/utils/dnd.ts
@@ -1,0 +1,10 @@
+export const draggableItem = {
+  card: 'card',
+};
+
+export type DragItem = {
+  type: string;
+  index: number;
+  listIndex: number;
+  id: string;
+};
